requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
cryptography==37.0.4
cssselect==1.1.0
cssutils==2.5.1
Django==3.1.7
django-filter==2.4.0
gevent==20.4.0
greenlet==0.4.15
gunicorn==20.0.4
hiredis==2.0.0
idna==2.10
importlib-metadata==4.12.0
PyJWT==1.6.4
lxml==4.9.1
mysqlclient==1.4.6
phonenumbers==8.12.25
premailer==3.10.0
pycparser==2.21
pydash==4.9.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==3.4.1
requests==2.23.0
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.25.11
yagmail==0.14.245
zipp==3.8.1